Module measure_time::kv

UNSTABLE: Structured key-value pairs.

This module is unstable and breaking changes may be made at any time. See the tracking issue for more details.

Add the kv_unstable feature to your Cargo.toml to enable this module:

[dependencies.log]
features = ["kv_unstable"]

Modules

source

Sources for key-value pairs.

value

Structured values.

Structs

Error

An error encountered while working with structured data.

Key

A key in a structured key-value pair.

Value

A value in a structured key-value pair.

Traits

Source

A source of key-value pairs.

ToKey

A type that can be converted into a Key.

ToValue

A type that can be converted into a Value.

Visitor

A visitor for the key-value pairs in a Source.